refactor: rewire backend entry point, delete old structure
- Create src/app.js: wires express, cors, cookieParser, routes, errorHandler
- Simplify index.js: dotenv → validateEnv → app.listen only
- Delete legacy files: supabase.js (root), src/auth/, src/projects/, src/users/, src/middleware/
- Backend now loads exclusively from the new modular structure
