--- project/node_runtime/src/users/users.routes.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const { getUsers, updateStatus, updateRole } = require('./users.controller');
-
-router.get('/', getUsers);
-router.put('/status/:id', updateStatus);
-router.put('/role/:id', updateRole);
-
-module.exports = router;
