--- project/node_runtime/supabase.js
@@ -1,10 +1,0 @@
-const {createClient} = require('@supabase/supabase-js')
-
-require('dotenv').config()
-
-const supabase = createClient(
-    process.env.SUPABASE_URL,
-    process.env.SUPABASE_KEY
-)
-
-module.exports = supabase
